requirements.txt
torchmetrics
segmentation_models_pytorch
pytorch_pretrained_vit
transformers
matplotlib 
jupyter
scipy 
escnn
timm
pandas
tqdm 
ultralytics-thop
fvcore
prettytable 
pyyaml 
omegaconf
albumentations==1.4.0 
opencv-python 
opencv-contrib-python
wandb
